requirements.txt
attrs==25.3.0
boto3==1.38.23
botocore==1.38.23
cattrs==24.1.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.2
flatbuffers==25.2.10
greenlet==3.2.2
h11==0.16.0
idna==3.10
jh2==5.0.9
jmespath==1.0.1
niquests==3.14.1
numpy==2.2.6
openmeteo_requests==1.5.0
openmeteo_sdk==1.20.0
pandas==2.2.3
platformdirs==4.3.8
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
qh3==1.5.1
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
s3transfer==0.13.0
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.4.0
urllib3-future==2.12.922
wassima==1.2.2
python-dotenv